refractor: Seperated Controls and Video
The controls and the video component are now two differently
working things with some sharing properties.

All the uneccessary things from the Video component are now
removed.
