+++ package.json
@@ -6,7 +6,8 @@
     "dev": "next dev",
     "build": "next build",
     "start": "next start",
-    "lint": "next lint"
+    "lint": "next lint",
+    "format": "npx prettier --write src/**/*.{js,ts,jsx,tsx,json,css,md}"
   },
   "dependencies": {
     "@svgr/webpack": "^8.1.0",
@@ -20,6 +21,7 @@
     "@types/react-dom": "^18",
     "eslint": "^8",
     "eslint-config-next": "14.2.3",
+    "prettier": "^3.3.2",
     "typescript": "^5"
   }
 }
