+++ src/app/page.tsx
@@ -2,7 +2,5 @@ import styles from "./page.module.css";
 import VideoPlayer from "@/components/VideoPlayer/index";
 
 export default function Home() {
-  return (
-    <VideoPlayer />
-  );
+  return <VideoPlayer />;
 }
