+++ src/components/FullScreenBtn/index.tsx
@@ -1,6 +1,5 @@
-
 import FullScreen from "../../../public/images/full-screen.svg";
-import CollapseScreen from "../../../public/images/collapse-screen.svg"
+import CollapseScreen from "../../../public/images/collapse-screen.svg";
 
 import styles from "@/components/VideoPlayer/VideoPlayer.module.css";
 
@@ -8,16 +7,16 @@ interface fullScreenProps {
   isFullScreen: boolean;
   onExpandCollapse: () => void;
 }
-const FullScreenBtn = ({ isFullScreen, onExpandCollapse } : fullScreenProps): JSX.Element => {
+const FullScreenBtn = ({
+  isFullScreen,
+  onExpandCollapse,
+}: fullScreenProps): JSX.Element => {
   return (
     <button
       className={styles.controlBtn}
       id={styles.fullScreen}
-      onClick={() => onExpandCollapse()}
-    >
-    {
-      isFullScreen ? <CollapseScreen /> : <FullScreen />
-    }
+      onClick={() => onExpandCollapse()}>
+      {isFullScreen ? <CollapseScreen /> : <FullScreen />}
     </button>
   );
 };
